组件目录结构改变。把studio library 下面目录移动到 components
