+++ dashboard/src/components/dict/Confidence.tsx
