+++ dashboard/src/components/dict/SelectCase.tsx
