+++ dashboard/src/components/template/Term.tsx
@@ -2,7 +2,7 @@ import { ProCard } from "@ant-design/pro-components";
 import { Button, Popover } from "antd";
 import { SearchOutlined } from "@ant-design/icons";
 import { Typography } from "antd";
-import TermCreate, { IWidgetDictCreate } from "../studio/term/TermCreate";
+import TermCreate, { IWidgetDictCreate } from "../term/TermCreate";
 import { command } from "../../reducers/command";
 import store from "../../store";
