+++ dashboard/src/pages/library/blog/course.tsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import BlogNav from "../../../components/library/blog/BlogNav";
+import BlogNav from "../../../components/blog/BlogNav";
 
 const Widget = () => {
 	// TODO
