+++ dashboard/src/pages/library/blog/term.tsx
@@ -1,15 +1,15 @@
 import { useParams } from "react-router-dom";
-import BlogNav from "../../../components/library/blog/BlogNav";
+import BlogNav from "../../../components/blog/BlogNav";
 
 const Widget = () => {
-	// TODO
-	const { studio } = useParams(); //url 参数
+  // TODO
+  const { studio } = useParams(); //url 参数
 
-	return (
-		<>
-			<BlogNav selectedKey="term" studio={studio ? studio : ""} />
-		</>
-	);
+  return (
+    <>
+      <BlogNav selectedKey="term" studio={studio ? studio : ""} />
+    </>
+  );
 };
 
 export default Widget;
