+++ dashboard/src/pages/studio/anthology/list.tsx
@@ -8,7 +8,7 @@ import type { MenuProps } from "antd";
 import { Button, Dropdown, Menu, Popover } from "antd";
 import { SearchOutlined } from "@ant-design/icons";
 
-import AnthologyCreate from "../../../components/studio/anthology/AnthologyCreate";
+import AnthologyCreate from "../../../components/anthology/AnthologyCreate";
 import { IAnthologyListResponse } from "../../../components/api/Article";
 import { get } from "../../../request";
 import { PublicityValueEnum } from "../../../components/studio/table";
@@ -16,220 +16,209 @@ import { PublicityValueEnum } from "../../../components/studio/table";
 const { Text } = Typography;
 
 const onMenuClick: MenuProps["onClick"] = (e) => {
-	console.log("click", e);
+  console.log("click", e);
 };
 
 const menu = (
-	<Menu
-		onClick={onMenuClick}
-		items={[
-			{
-				key: "1",
-				label: "在藏经阁中打开",
-				icon: <SearchOutlined />,
-			},
-			{
-				key: "2",
-				label: "分享",
-				icon: <SearchOutlined />,
-			},
-		]}
-	/>
+  <Menu
+    onClick={onMenuClick}
+    items={[
+      {
+        key: "1",
+        label: "在藏经阁中打开",
+        icon: <SearchOutlined />,
+      },
+      {
+        key: "2",
+        label: "分享",
+        icon: <SearchOutlined />,
+      },
+    ]}
+  />
 );
 
 interface IItem {
-	sn: number;
-	id: string;
-	title: string;
-	subtitle: string;
-	publicity: number;
-	articles: number;
-	createdAt: number;
+  sn: number;
+  id: string;
+  title: string;
+  subtitle: string;
+  publicity: number;
+  articles: number;
+  createdAt: number;
 }
 
 const Widget = () => {
-	const intl = useIntl();
-	const { studioname } = useParams();
-	const anthologyCreate = <AnthologyCreate studio={studioname} />;
-	return (
-		<>
-			<ProTable<IItem>
-				columns={[
-					{
-						title: intl.formatMessage({
-							id: "dict.fields.sn.label",
-						}),
-						dataIndex: "sn",
-						key: "sn",
-						width: 50,
-						search: false,
-					},
-					{
-						title: intl.formatMessage({
-							id: "forms.fields.title.label",
-						}),
-						dataIndex: "title",
-						key: "title",
-						tip: "过长会自动收缩",
-						ellipsis: true,
-						render: (text, row, index, action) => {
-							return (
-								<div>
-									<div>
-										<Link
-											to={`/studio/${studioname}/anthology/${row.id}/edit`}
-										>
-											{row.title}
-										</Link>
-									</div>
-									<Text type="secondary">{row.subtitle}</Text>
-								</div>
-							);
-						},
-					},
-					{
-						title: intl.formatMessage({
-							id: "forms.fields.publicity.label",
-						}),
-						dataIndex: "publicity",
-						key: "publicity",
-						width: 100,
-						search: false,
-						filters: true,
-						onFilter: true,
-						valueEnum: PublicityValueEnum(),
-					},
-					{
-						title: intl.formatMessage({
-							id: "article.fields.article.count.label",
-						}),
-						dataIndex: "articles",
-						key: "articles",
-						width: 100,
-						search: false,
-						sorter: (a, b) => a.articles - b.articles,
-					},
-					{
-						title: intl.formatMessage({
-							id: "forms.fields.created-at.label",
-						}),
-						key: "created-at",
-						width: 100,
-						search: false,
-						dataIndex: "createdAt",
-						valueType: "date",
-						sorter: (a, b) => a.createdAt - b.createdAt,
-					},
-					{
-						title: intl.formatMessage({ id: "buttons.option" }),
-						key: "option",
-						width: 120,
-						valueType: "option",
-						render: (text, row, index, action) => [
-							<Dropdown.Button
-								type="link"
-								key={index}
-								overlay={menu}
-							>
-								{intl.formatMessage({
-									id: "buttons.edit",
-								})}
-							</Dropdown.Button>,
-						],
-					},
-				]}
-				rowSelection={{
-					// 自定义选择项参考: https://ant.design/components/table-cn/#components-table-demo-row-selection-custom
-					// 注释该行则默认不显示下拉选项
-					selections: [Table.SELECTION_ALL, Table.SELECTION_INVERT],
-				}}
-				tableAlertRender={({
-					selectedRowKeys,
-					selectedRows,
-					onCleanSelected,
-				}) => (
-					<Space size={24}>
-						<span>
-							{intl.formatMessage({ id: "buttons.selected" })}{" "}
-							{selectedRowKeys.length}
-							<Button
-								type="link"
-								style={{ marginInlineStart: 8 }}
-								onClick={onCleanSelected}
-							>
-								{intl.formatMessage({ id: "buttons.unselect" })}
-							</Button>
-						</span>
-					</Space>
-				)}
-				tableAlertOptionRender={() => {
-					return (
-						<Space size={16}>
-							<Button type="link">批量删除</Button>
-						</Space>
-					);
-				}}
-				request={async (params = {}, sorter, filter) => {
-					// TODO
-					console.log(params, sorter, filter);
-					let url = `/v2/anthology?view=studio&name=${studioname}`;
-					const offset =
-						((params.current ? params.current : 1) - 1) *
-						(params.pageSize ? params.pageSize : 20);
-					url += `&limit=${params.pageSize}&offset=${offset}`;
-					if (typeof params.keyword !== "undefined") {
-						url +=
-							"&search=" + (params.keyword ? params.keyword : "");
-					}
+  const intl = useIntl();
+  const { studioname } = useParams();
+  const anthologyCreate = <AnthologyCreate studio={studioname} />;
+  return (
+    <>
+      <ProTable<IItem>
+        columns={[
+          {
+            title: intl.formatMessage({
+              id: "dict.fields.sn.label",
+            }),
+            dataIndex: "sn",
+            key: "sn",
+            width: 50,
+            search: false,
+          },
+          {
+            title: intl.formatMessage({
+              id: "forms.fields.title.label",
+            }),
+            dataIndex: "title",
+            key: "title",
+            tip: "过长会自动收缩",
+            ellipsis: true,
+            render: (text, row, index, action) => {
+              return (
+                <div>
+                  <div>
+                    <Link to={`/studio/${studioname}/anthology/${row.id}/edit`}>
+                      {row.title}
+                    </Link>
+                  </div>
+                  <Text type="secondary">{row.subtitle}</Text>
+                </div>
+              );
+            },
+          },
+          {
+            title: intl.formatMessage({
+              id: "forms.fields.publicity.label",
+            }),
+            dataIndex: "publicity",
+            key: "publicity",
+            width: 100,
+            search: false,
+            filters: true,
+            onFilter: true,
+            valueEnum: PublicityValueEnum(),
+          },
+          {
+            title: intl.formatMessage({
+              id: "article.fields.article.count.label",
+            }),
+            dataIndex: "articles",
+            key: "articles",
+            width: 100,
+            search: false,
+            sorter: (a, b) => a.articles - b.articles,
+          },
+          {
+            title: intl.formatMessage({
+              id: "forms.fields.created-at.label",
+            }),
+            key: "created-at",
+            width: 100,
+            search: false,
+            dataIndex: "createdAt",
+            valueType: "date",
+            sorter: (a, b) => a.createdAt - b.createdAt,
+          },
+          {
+            title: intl.formatMessage({ id: "buttons.option" }),
+            key: "option",
+            width: 120,
+            valueType: "option",
+            render: (text, row, index, action) => [
+              <Dropdown.Button type="link" key={index} overlay={menu}>
+                {intl.formatMessage({
+                  id: "buttons.edit",
+                })}
+              </Dropdown.Button>,
+            ],
+          },
+        ]}
+        rowSelection={{
+          // 自定义选择项参考: https://ant.design/components/table-cn/#components-table-demo-row-selection-custom
+          // 注释该行则默认不显示下拉选项
+          selections: [Table.SELECTION_ALL, Table.SELECTION_INVERT],
+        }}
+        tableAlertRender={({
+          selectedRowKeys,
+          selectedRows,
+          onCleanSelected,
+        }) => (
+          <Space size={24}>
+            <span>
+              {intl.formatMessage({ id: "buttons.selected" })}{" "}
+              {selectedRowKeys.length}
+              <Button
+                type="link"
+                style={{ marginInlineStart: 8 }}
+                onClick={onCleanSelected}
+              >
+                {intl.formatMessage({ id: "buttons.unselect" })}
+              </Button>
+            </span>
+          </Space>
+        )}
+        tableAlertOptionRender={() => {
+          return (
+            <Space size={16}>
+              <Button type="link">批量删除</Button>
+            </Space>
+          );
+        }}
+        request={async (params = {}, sorter, filter) => {
+          // TODO
+          console.log(params, sorter, filter);
+          let url = `/v2/anthology?view=studio&name=${studioname}`;
+          const offset =
+            ((params.current ? params.current : 1) - 1) *
+            (params.pageSize ? params.pageSize : 20);
+          url += `&limit=${params.pageSize}&offset=${offset}`;
+          if (typeof params.keyword !== "undefined") {
+            url += "&search=" + (params.keyword ? params.keyword : "");
+          }
 
-					const res = await get<IAnthologyListResponse>(url);
-					const items: IItem[] = res.data.rows.map((item, id) => {
-						const date = new Date(item.created_at);
-						return {
-							sn: id + 1,
-							id: item.uid,
-							title: item.title,
-							subtitle: item.subtitle,
-							publicity: item.status,
-							articles: item.childrenNumber,
-							createdAt: date.getTime(),
-						};
-					});
-					console.log(items);
-					return {
-						total: res.data.count,
-						succcess: true,
-						data: items,
-					};
-				}}
-				rowKey="id"
-				bordered
-				pagination={{
-					showQuickJumper: true,
-					showSizeChanger: true,
-				}}
-				search={false}
-				options={{
-					search: true,
-				}}
-				toolBarRender={() => [
-					<Popover
-						content={anthologyCreate}
-						title="new article"
-						placement="bottomRight"
-					>
-						<Button
-							key="button"
-							icon={<PlusOutlined />}
-							type="primary"
-						>
-							{intl.formatMessage({ id: "buttons.create" })}
-						</Button>
-					</Popover>,
-				]}
-			/>
-		</>
-	);
+          const res = await get<IAnthologyListResponse>(url);
+          const items: IItem[] = res.data.rows.map((item, id) => {
+            const date = new Date(item.created_at);
+            return {
+              sn: id + 1,
+              id: item.uid,
+              title: item.title,
+              subtitle: item.subtitle,
+              publicity: item.status,
+              articles: item.childrenNumber,
+              createdAt: date.getTime(),
+            };
+          });
+          console.log(items);
+          return {
+            total: res.data.count,
+            succcess: true,
+            data: items,
+          };
+        }}
+        rowKey="id"
+        bordered
+        pagination={{
+          showQuickJumper: true,
+          showSizeChanger: true,
+        }}
+        search={false}
+        options={{
+          search: true,
+        }}
+        toolBarRender={() => [
+          <Popover
+            content={anthologyCreate}
+            title="new article"
+            placement="bottomRight"
+          >
+            <Button key="button" icon={<PlusOutlined />} type="primary">
+              {intl.formatMessage({ id: "buttons.create" })}
+            </Button>
+          </Popover>,
+        ]}
+      />
+    </>
+  );
 };
 
 export default Widget;
